Added API call functionality from client
With this update, the client actually does receive a response from the server. That being said, there is also some work to be done in terms of streamlining, because right now every time a user sends a job description, they also resend the document. This is obviously to be amended because there will be individual users and it won't just be document upload, but this is something that ought to be fixed.